+++ minijinja/src/vm/context.rs
@@ -129,6 +129,7 @@ impl From<Vec<Value>> for Stack {
 }
 
 pub(crate) struct Context<'env> {
+    env: &'env Environment<'env>,
     stack: Vec<Frame<'env>>,
     outer_stack_depth: usize,
     recursion_limit: usize,
@@ -136,62 +137,31 @@ pub(crate) struct Context<'env> {
 
 impl fmt::Debug for Context<'_> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        fn dump<'a>(
-            m: &mut std::fmt::DebugMap,
-            seen: &mut HashSet<Cow<'a, str>>,
-            ctx: &'a Context<'a>,
-        ) -> fmt::Result {
-            for frame in ctx.stack.iter().rev() {
-                for (key, value) in frame.locals.iter() {
-                    if !seen.contains(&Cow::Borrowed(*key)) {
-                        m.entry(&key, value);
-                        seen.insert(Cow::Borrowed(key));
-                    }
-                }
-
-                if let Some(ref l) = frame.current_loop {
-                    if l.with_loop_var && !seen.contains("loop") {
-                        m.entry(&"loop", &l.object);
-                        seen.insert(Cow::Borrowed("loop"));
-                    }
-                }
-
-                if let Ok(iter) = frame.ctx.try_iter() {
-                    for key in iter {
-                        if let Some(str_key) = key.as_str() {
-                            if !seen.contains(&Cow::Borrowed(str_key)) {
-                                if let Ok(value) = frame.ctx.get_item(&key) {
-                                    m.entry(&str_key, &value);
-                                    seen.insert(Cow::Owned(str_key.to_owned()));
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            Ok(())
-        }
-
-        let mut m = f.debug_map();
-        let mut seen = HashSet::new();
-        ok!(dump(&mut m, &mut seen, self));
-        m.finish()
+        let mut vars = Vec::from_iter(self.known_variables(false));
+        vars.sort();
+        f.debug_map()
+            .entries(vars.into_iter().map(|key| {
+                let value = self.load(&key).unwrap_or_default();
+                (key, value)
+            }))
+            .finish()
     }
 }
 
 impl<'env> Context<'env> {
     /// Creates an empty context.
-    pub fn new(recursion_limit: usize) -> Context<'env> {
+    pub fn new(env: &'env Environment<'env>) -> Context<'env> {
         Context {
+            env,
             stack: Vec::with_capacity(32),
             outer_stack_depth: 0,
-            recursion_limit,
+            recursion_limit: env.recursion_limit(),
         }
     }
 
     /// Creates a context
-    pub fn new_with_frame(frame: Frame<'env>, recursion_limit: usize) -> Context<'env> {
-        let mut rv = Context::new(recursion_limit);
+    pub fn new_with_frame(env: &'env Environment<'env>, frame: Frame<'env>) -> Context<'env> {
+        let mut rv = Context::new(env);
         rv.stack.push(frame);
         rv
     }
@@ -214,7 +184,7 @@ impl<'env> Context<'env> {
     /// to emulate the behavior of how scopes work in Jinja2 in Python.  The
     /// unfortunate downside is that this has to be done with a `Mutex`.
     #[cfg(feature = "macros")]
-    pub fn enclose(&mut self, env: &Environment, key: &str) {
+    pub fn enclose(&mut self, key: &str) {
         self.stack
             .last_mut()
             .unwrap()
@@ -222,7 +192,7 @@ impl<'env> Context<'env> {
             .as_mut()
             .unwrap()
             .clone()
-            .store_if_missing(key, || self.load(env, key).unwrap_or(Value::UNDEFINED));
+            .store_if_missing(key, || self.load(key).unwrap_or(Value::UNDEFINED));
     }
 
     /// Loads the closure and returns it.
@@ -261,7 +231,7 @@ impl<'env> Context<'env> {
     }
 
     /// Looks up a variable in the context.
-    pub fn load(&self, env: &Environment, key: &str) -> Option<Value> {
+    pub fn load(&self, key: &str) -> Option<Value> {
         for frame in self.stack.iter().rev() {
             // look at locals first
             if let Some(value) = frame.locals.get(key) {
@@ -282,7 +252,41 @@ impl<'env> Context<'env> {
             }
         }
 
-        env.get_global(key)
+        self.env.get_global(key)
+    }
+
+    /// Returns an iterable of all declared variables.
+    pub fn known_variables(&self, with_globals: bool) -> HashSet<Cow<'_, str>> {
+        let mut seen = HashSet::<Cow<'_, str>>::new();
+        for frame in self.stack.iter().rev() {
+            for key in frame.locals.keys() {
+                if !seen.contains(&Cow::Borrowed(*key)) {
+                    seen.insert(Cow::Borrowed(key));
+                }
+            }
+
+            if let Some(ref l) = frame.current_loop {
+                if l.with_loop_var && !seen.contains("loop") {
+                    seen.insert(Cow::Borrowed("loop"));
+                }
+            }
+
+            if let Ok(iter) = frame.ctx.try_iter() {
+                for key in iter {
+                    if let Some(str_key) = key.as_str() {
+                        if !seen.contains(&Cow::Borrowed(str_key))
+                            && frame.ctx.get_item(&key).is_ok()
+                        {
+                            seen.insert(Cow::Owned(str_key.to_owned()));
+                        }
+                    }
+                }
+            }
+        }
+        if with_globals {
+            seen.extend(self.env.globals().map(|x| Cow::Borrowed(x.0)));
+        }
+        seen
     }
 
     /// Pushes a new layer.